+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changes
 
+- Update video block to the sidebar and get the image block look and feel @sneridagh
+
 ### Internal
 
 - Fix and update to latest husky and lint-staged @sneridagh